RUB-287 rust: split connect block digest tests
